Move type traits into independent header
