+++ include/RAJA/policy/hip/forall.hpp
@@ -532,8 +532,7 @@ forall_impl(resources::Hip hip_res,
   Iterator begin = std::begin(iter);
   Iterator end   = std::end(iter);
   IndexType len  = std::distance(begin, end);
-  constexpr bool is_forallparampack_empty =
-      RAJA::expt::type_traits::is_ForallParamPack_empty<ForallParam>::value;
+
   // Only launch kernel if we have something to iterate over
   if (len > 0)
   {
